[Experiments] montessori demo is on its way

+++ semantic_digital_twin/src/semantic_digital_twin/adapters/urdf.py
@@ -273,7 +273,7 @@ class URDFParser:
             parent_T_connection_expression=parent_T_connection,
             multiplier=multiplier,
             offset=offset,
-            axis=Vector3(*map(int, joint.axis), reference_frame=parent),
+            axis=Vector3(*map(float, joint.axis), reference_frame=parent),
             raw_dof=dof,
         )
         return result
